DYI: Move common validators into mixin/preference-validators.js
